Remove `Listbox` (#2142)
# Pull Request

## 🤨 Rationale

The `Listbox` component was initially implemented with the idea that it
would be used as part of the `RichTextEditor`. However, we pivoted away
from that notion and the `Listbox` is not used anywhere (and is not a
documented component). Instead of making the necessary changes to allow
it to behave properly with not changes that are being introduced it
makes more sense to just remove it for now.

## 👩‍💻 Implementation

Mainly deleting files. Updated a template file to point to FAST export
instead of something from the now deleted `Listbox` file.

## 🧪 Testing

N/A

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.