Refactored wafer map experimental structure and update cycle  (#2042)
# Pull Request

## 🤨 Rationale

The experimental component part needed to be more centralized with it's
internal state and configuration.

The update needed to consider the new asynchronous calls and preserving
the state between subsequent calls.

## 👩‍💻 Implementation

Centralized all the state management under computations.
Created a snapshot mechanism to preserve data state between async calls.
Combined all the worker configurations in two objects to make a minimal
number of async calls.
 
## 🧪 Testing

merged most of the tests in the computations spec. 

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.

## Open Issues

- worker type export to the wafer. right now, I copy and paste the file
contents
- better structure/ update cycle?

---------

Co-authored-by: Milan Raj <rajsite@users.noreply.github.com>