Disable three more intermittent rich text tests (#2160)
# Pull Request

## 🤨 Rationale

Seems to be the same issue that caused Jesse to recently disable three
RichTextEditorMention tests. See #2150. There are probably other tests
that are intermittent, but I don't want to pre-emptively disable tests
until we see them fail.

## 👩‍💻 Implementation

The usual.

## 🧪 Testing

N/A

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: Milan Raj <rajsite@users.noreply.github.com>