Address issue with select not recognizing when group options are updated. (#2162)
# Pull Request

## 🤨 Rationale

Realized from a conversation that the `Select` wasn't reacting to
updates to `ListOptionGroup` slotted content updates.

## 👩‍💻 Implementation

Simply needed to have the `Select` observe the `listOption` property of
the `ListOptionGroup` elements it owns, and then call
`slottedOptionsChanged` directly. Also required a slight change to how
we maintain the currently selected option when the options are changed,
particularly for when the selected option is removed.

## 🧪 Testing

Added a few unit tests.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.