Remove StoryBlocks from linting (#2187)
# Pull Request

## 🤨 Rationale

After #2168 the storybook build became intermittent. The only changes in
the storybook package that seem like it could be build related was the
addition of a `tsconfig` to handle the `tsx` file in the `.storybook`
config directory which is outside of the `src` directory. Fixes #2179

## 👩‍💻 Implementation

Remove the additional `tsconfig` files.

## 🧪 Testing

Ran several clean builds locally, hopefully resolves the issue, but can
re-open if still comes up.

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.