Remove node polyfills no longer needed for rich text components (#2195)
# Pull Request

## 🤨 Rationale

Fixes #1700 These two polyfills were added in #1335 because markdown-it
implicitly depended on some functionality that existed in Node but not
browser environments. We've since upgraded markdown-it to a version that
expresses those dependencies on normal npm packages, so the polyfills
are no longer needed.
https://github.com/markdown-it/markdown-it/issues/967

## 👩‍💻 Implementation

Remove rollup plugins and npm dependencies on those plugins that were
added in #1335.

## 🧪 Testing

Mostly depending on PR build but also spot checking rich text
functionality in storybook.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.