List option in group needs to register with Select (#2198)
# Pull Request

## 🤨 Rationale

Had a realization that list options that were in a group would not go
through the `registerOption` routine (a fix for SLE Angular apps). While
I'm not sure if there is any current use-case that would hit the
problem, I went ahead and made the fix and the tests that demonstrate
the problem.