Update `SelectPageObject` `clickOption` to only deal with visible options. (#2221)
# Pull Request

## 🤨 Rationale

While doing the integration work to begin leveraging a select with
grouping in SLE, I recognized that an existing page object had an API
for selecting options by index, where the index was related to only the
visible options in the dropdown. I feel this is more intuitive for a
pageObject API than what we currently had in our own, so I'm updating
our pageObject implementation to follow this.

## 👩‍💻 Implementation

Simply finding the visible options with which to relate the provided
`index` parameter to `clickOption` against.

## 🧪 Testing

Updated tests that had to adjust the index they were using to align with
only the visible options.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.