Fix table phantom multi-select (#2242)
# Pull Request

## 🤨 Rationale

Fixes #2222

Because the event emitted by the row (`row-selection-toggle`) does not
include details about whether shift (or any other modifier key) was
pressed as part of the selection toggle--and it can't, because the
checkbox change event doesn't know--the table has to track the state of
the shift key by listening for `keyup` and `keydown` events on the
document. This breaks down if the shift key is let go while focus is
outside the document (e.g. on the browser chrome). At that point, the
table still thinks the shift key is depressed and behaves accordingly.

## 👩‍💻 Implementation

We were previously only updating the table's `documentShiftKeyDown` flag
when the `shift` key was pressed/unpressed. But we can also look at the
`shiftKey` flag of the event data for other keys' events. For example,
if the `space` key is pressed to select a row, and `event.shiftKey` is
`false`, then we can update `documentShiftKeyDown` to prevent the
incorrect behavior.

However, we still have to handle the case where the selection is made by
mouse click rather than pressing `space`. In that scenario, we may not
get any `keydown`/`keyup` events that would clue us in that `shift` is
no longer held down. My solution is to add a `blur` handler on the
window that clears `documentShiftKeyDown` when the focus leaves the
window. It would be better if we could detect focus leaving the document
rather than window, but the document doesn't get a `blur` event, and
`focusout` bubbles up from all elements in a document. Generally, if the
window loses focus, we can expect that the user is no longer in the
process of making a selection. However there are contrived use cases
where this can result in the wrong behavior. For example:
1. an element in the window has focus
2. hold down `shift`
3. press `tab` repeatedly (`shift` is still pressed) until focus leaves
the window, cycles through everything else, and returns to the table
4. click to select a row. The table will behave as if `shift` is not
pressed, even though it is.

Note that in step 4, if `space` is pressed rather than mouse-clicking,
the `keydown` for `space` will update `documentShiftKeyDown`, and things
will behave properly.

## 🧪 Testing

Wrote a couple unit tests and manually tested in Storybook.

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.