Allow text column and anchor column to be sorted by custom data field (#2224)
# Pull Request

## 🤨 Rationale

Add ability to sort a text column or anchor column by a custom data
field. This implements the capability described in #2206.

This PR only includes the nimble-components changes. Angular and Blazor
integration will be done in a separate PR.

## 👩‍💻 Implementation

- Create `mixinCustomSortOrderColumnAPI` that does the following for
columns using it
    - Adds a `sort-by-field-name` attribute
    - Requires a validator with the `invalidCustomSortWithGrouping` flag
- Updates the validator's `invalidCustomSortWithGrouping` flag based on
the values of `sortByFieldName`, `columnInternals.groupIndex`, and
`columnInternals.groupingDisabled`
- Use the mixin on the text column and anchor column
- Update storybook stories for text column and anchor column

## 🧪 Testing

- New unit tests for mixin
- Manually tested in storybook

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.