Add combobox page object and clean up combobox tests (#2255)
# Pull Request

## 🤨 Rationale

A page object is useful for writing clearer, less error-prone unit
tests. I was going to write new combobox tests and was motivated to
follow the example of the select unit tests.

## 👩‍💻 Implementation

Based some of the page object on the select page object, but removed all
functions that are not currently needed. We can easily add additional
functions as needed.

Tests were adapted to use the page object. One duplicate test
(copy-paste error) was removed. `parameterizeSpec` was employed in two
places.

## 🧪 Testing

Unit tests pass.

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: Milan Raj <rajsite@users.noreply.github.com>