Consolidate table slot tracking (#2271)
# Pull Request

## 🤨 Rationale

Previously there were two different ways that client-requested slots
were tracked and dynamically created within rows of the table. This PR
consolidates that code so that action menu slots and cell-requested
slots all end up in `columnRequestedSlots`, and ultimately as part of
`tableData`.

Resolves #2248 

## 👩‍💻 Implementation

When determining that action menu slots need to be created, the table
now creates those slots in `columnRequestedSlots` rather than simply
ensuring that slots with the same name don't already exist in
`columnRequestedSlots`. The code that handles action menu slot names
changing was also updated to remove the old slots from
`columnRequestedSlots` and to add the new slots to
`columnRequestedSlots` when the slot names change while the menu is
open.

## 🧪 Testing

- Sanity checked that action menus still work as expected in storybook
- Verified existing unit tests still pass

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: Milan Raj <rajsite@users.noreply.github.com>