Remove reserved space for the table's "Collapse all" button when the table cannot have expandable rows (#2261)
# Pull Request

## 🤨 Rationale

When all table columns have grouping disabled and the table isn't
configured for hierarchical data, then there isn't a way for a row to be
collapsable. Therefore, the "Collapse all" button will never be shown,
and the space for it does not need to be reserved.

When the space is not reserved for the "Collapse all" button, there now
is a 16px margin between the edge of the table (or selection checkbox
when `multiple` selection is enabled), rather than an 8px margin. This
spacing is specified in the visual designs.

Resolves #2241 

## 👩‍💻 Implementation

The table checks to see if any column supports grouping or if
`parentIdFieldName` is configured. If so, the space is still reserved
for the "Collapse all" button. If not, space is not reserved in the
header for the "Collapse all" button, and the table tells the row not to
reserve space for an expand/collapse button.

## 🧪 Testing

- Manually tested in storybook, including resizing columns
- Added new matrix tests
- Added new unit tests

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.