Refinements to table selection when using SHIFT (#2254)
# Pull Request

## 🤨 Rationale

This PR addresses two shift-selection concerns:
1. Range selection using shift should not start on a row that is not
selected
2. Shift+click to a row should deselect any rows not in the selected
range, but Shift+Ctrl+click should allow a range to be added to the
current select. (Note, this behavior is specific to clicking on the rows
themselves in multi-select mode. Clicking on a selection checkbox will
always behave as though Ctrl is held.)

Resolves #2239

## 👩‍💻 Implementation

The desired multiselect behavior is that a row will only be considered
the start of a shift-select range if that row was being selected (either
via selection checkbox or via clicking the row), but it will not be
considered the start of a shift-select range if that row was deselected.

To achieve this behavior, I've updated the `MultiSelectionManager` to
assign the current row ID to `shiftSelectStartRowId` only if the row is
being selected. If the row is being deselected, `shiftSelectStartRowId`
is reset to `undefined`.

## 🧪 Testing

- Manually tested in storybook, including keyboard selection
- Added new unit tests

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: m-akinc <7282195+m-akinc@users.noreply.github.com>
Co-authored-by: Milan Raj <rajsite@users.noreply.github.com>