Expose combobox page object from nimble-angular (#2286)
# Pull Request

## 🤨 Rationale

We added a combobox page object in `nimble-components` in #2255. I would
like to use it from an SLE test but it's not exposed from
`nimble-angular`

## 👩‍💻 Implementation

Copy the select page object angular entry point and rename it to
combobox.

## 🧪 Testing

I'll use this in a follow up PR in the SLE repo.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: Milan Raj <rajsite@users.noreply.github.com>