Add new SelectPageObject API for setting filter text. (#2292)
# Pull Request

## 🤨 Rationale

While trying to refactor the SLE user selector to use the
`nimble-select`, I ran into some testing roadblocks that were reliant on
`fakieAsync` for debouncing purposes. At present the `SelectPageObject`
API offered no non-async way for setting the filter text, which is
needed when in a `fakeAsync` context. This PR provides the means for
doing this.

## 👩‍💻 Implementation

Simple API for setting the filter text.

## 🧪 Testing

Added a pageObject test that exercises the implementation.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: Milan Raj <rajsite@users.noreply.github.com>