Add dropdown "No results" placeholder for combobox and rich-text mention listbox (#2291)
# Pull Request

## 🤨 Rationale

Part of #1309

The other part of that issue is updating clients to use this instead of
their workarounds.

The issue specifies that the text should be configurable. This can be
done by way of a `nimble-label-provider-core` element, but that has
limitations, i.e. if you want to customize two different
combobox/select/rich-text-listboxes on the same page, they have to be
scoped to distinct theme provider elements.

## 👩‍💻 Implementation

- added support for "No results found" placeholder to the combobox and
rich-text mention listbox
- moved related styling to shared dropdown styling
   - reordered to follow DOM order
- fixed issue with combobox where options with `hidden` attribute would
still be included in `filteredOptions` (preventing placeholder from
showing)

## 🧪 Testing

- new unit tests for combobox and mention listbox
- new visual tests for combobox and mention listbox

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: Jonathan Meyer <26874831+atmgrifter00@users.noreply.github.com>
Co-authored-by: Milan Raj <rajsite@users.noreply.github.com>