Update TanStack and fix a row indentation issue (#2301)
# Pull Request

## 🤨 Rationale

Fixes #2289 
Fixes #2284

## 👩‍💻 Implementation

TanStack fixed the bug that was causing #2284, so updating TanStack
resolved that issue. However, some of the TanStack code changed
slightly, so I had to update our check for determining if a row was the
direct child of a group row in a table without hierarchy enabled.

## 🧪 Testing

- Ensured there were no chromatic diffs
- Manually tested that grouping and ungrouping a column with hierarchy
results in the correct indentation of rows

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.