Fix bug where user can't type disabled option into combobox (#2294)
# Pull Request

## 🤨 Rationale

Fixes #2252

## 👩‍💻 Implementation

The `options` getter should not revert to returning all the options
(i.e. `_options`) when `filteredOptions` is empty. Doing so can lead to
setting a `selectedIndex` that points to an option that is filtered out
or disabled.

The expected behavior now is that a disabled option functions exactly
like an enabled option, except that it cannot be clicked, cannot be
focused in the dropdown, and appears dimmed. Like enabled options, t can
be filtered to (if the input text matches it), and the user can set the
control's value to it by typing it in fully.

## 🧪 Testing

Added new tests

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.