Fix menu item indentation bug with dynamically slotted icons (#2296)
# Pull Request

## 🤨 Rationale

Fixes #2266

## 👩‍💻 Implementation

Proper indentation relies on `startColumnCount` being set on the menu
items. This is handled by
[`FoundationMenu.setItems`](https://github.com/microsoft/fast/blob/1f9c97df9acae24be44f239bd66b5b2157c7f3e5/packages/web-components/fast-foundation/src/menu/menu.ts#L253).
It needs to be called whenever the children of the slotted menu items
change. To do this, we need to modify the template so that it watches
for any added/removed icons that are children of its menu items (via
`children` directive). I've forked the FAST `Menu` class (into separate
file) and the template to modify them.

These modifications are in `menu.foundation.ts` rather than `index.ts`
because this is the kind of bug fix that we would have upstreamed to
FAST.

## 🧪 Testing

Added unit tests and forked FAST `Menu` tests.

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.