Update menu button column styles (#2309)
# Pull Request

## 🤨 Rationale

Resolves #2247

The only item from #2247 that we agreed should be changed is the height
of the menu button. It will no longer be styled with a height of
`controlSlimHeight` and instead have its default height. Additionally,
@fredvisser wanted the row's focus-visible outline to be centered
correctly when making this change since the misalignment of the outline
is significantly more apparent when there is a 32px control in a cell.
Therefore, I've also address that in this PR.

## 👩‍💻 Implementation

- Remove styling of the menu-button's height
- Update the focus-visible styling of a row so that the green outline
now symmetrically outlines the row. This now matches the outline height
of each cell. Previously the top border of the outline was 2px too high
because the row borders aren't symmetrical -- specifically, a row has a
2px top border and a 0px bottom border in order to have the horizontal
dividers between rows.

## 🧪 Testing

Manually tested

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: Milan Raj <rajsite@users.noreply.github.com>