Table data assignment bug (#2300)
# Pull Request

## 🤨 Rationale

Resolves #2265 

## 👩‍💻 Implementation

The `arrayToTree` code path for setting data in the table does not make
a copy of each record in the same way that the other `setData` code
paths do. Therefore, I updated the `arrayToTree` function to copy the
record as it turns the array into a tree.

## 🧪 Testing

- Added a new unit test that failed before my change and now passes
- Note, there already is a similar test for non-hierarchical data, so I
only added a new test for hierarchical data.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.