Updates to table page objects (#2308)
# Pull Request

## 🤨 Rationale

Update the table page object to include functionality needed by some
clients. Specifically,

- Add `getHeaderTextContent` to the page object so that clients don't
have to access the header element to get the text content themselves
- Export `TableColumnMappingPageObject` from Angular
- Extends `TablePageObject` that's exported from Angular to include a
new `waitForDataUpdatesToRender` to help clients wait for data to be
rendered in the table when they are using the `data$` property on the
Angular table directive, which makes the timing of `setData()` outside
of the client's control

## 👩‍💻 Implementation

Both the implementation of `getHeaderTextContent` and
`waitForDataUpdatesToRender` were copied from the `sl-table` page object
in SystemLinkShared.

## 🧪 Testing

- Updated a test in nimble-components to use the new nimble-components
page object test
- Wrote new Angular tests to test the Angular page object

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.