Fix Select arrow down and up behaviors with hidden selected option (#2314)
# Pull Request

## 🤨 Rationale

While doing the integration work of updating the SLE user selector to
use dynamic options, a non-ideal behavior became apparent in the case
where a user would provide some filter text, filtering out the current
selected user (which as our guidance states, requires the option to be
present but hidden), and then press `ArrowDown`, and nothing would get
selected. This would happen when the selected option was _below_ the
other options in DOM order.

The two solutions for creating the expected behavior were 1) update the
integration code to always put the hidden option at the top of the DOM
order, or 2) change the Nimble Select behavior to provide the expected
behavior in this case. I opted with 2 because it feels like a gap that
any user of the Nimble Select could possibly hit.

## 👩‍💻 Implementation

Now handling the `ArrowDown` and `ArrowUp` cases directly in the key
handling code of the `Select`.

## 🧪 Testing

Added some unit tests. Manually checked behavior in Angular example app
with dynamic options.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: Milan Raj <rajsite@users.noreply.github.com>