waitForEvent utility refactor (#2335)
# Pull Request

## 🤨 Rationale

The previous `createEventListener` helper in nimble testing utilities
leveraged a jasmine library feature which indirectly made jasmine a
dependency of nimble as several of the page objects had a dependency on
it and are part of the public api. This refactor breaks the jasmine
dependency in the public api.

## 👩‍💻 Implementation

- The `waitForEvent` helper was updated to take an optional event
handler callback. This deprecates the need for the `createEventListener`
utility which was removed.
- The tests were refactored to create their own jasmine spies instead of
relying on a utility to create them. This had several benefits:
- Many tests were using the utility that created spies unnecessarily.
They should have been using waitForEvent and now do so.
- Several tests where trying to track ordering of events. This is very
natural if the same spy is used in multiple event listeners which you
can now do since the spy is provided.
- Several tests were relying on the idea that the helpers can trigger
events multiple times. This is not the case, they are written to only
resolve once and to remove their event handlers after the first call
- Now that the spy is constructed in the test, it is easier to fulfill
it's generic and remove a bunch of `as` casting

## 🧪 Testing

Relying on existing test runs.

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.