Switch backdrop to using design tokens (#2339)
# Pull Request

## 🤨 Rationale

Fixes #1835 by using the existing `modalBackdropColor` for dialogs and
drawers now that CSS Custom Properties can participate in the DOM
cascade.

The related goal is to provide a workaround for #2304 to allow
temporarily overriding the backdrop color via CSS. For example:

```html
<nimble-dialog style="--ni-nimble-modal-backdrop-color: transparent;">
```

## 👩‍💻 Implementation

- Switch to the design token
- Drop the static token constants file
- I did not update the storybook docs to describe the workaround case or
add testing specific to it. I don't think we would describe overriding
the backdrop color as an explicitly supported configuration.

## 🧪 Testing

Relying on chromatic. Note, this actually fails to render the backdrop
on Chromatic. The behavior change was introduced in [Chrome
122](https://caniuse.com/mdn-css_selectors_backdrop_inherit_from_originating_element)
but Chromatic infrastructure is on [Chrome
121](https://www.chromatic.com/docs/infrastructure-release-notes/#version-6).

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.