Add additional theme aware tokens for calendar (#2352)
# Pull Request

## 🤨 Rationale

As per the UI mockup, adding additional theme-aware tokens for missing
parts in the calendar component.

Figma link to UI mockup (all three themes in different pages):
https://www.figma.com/design/VS7WavgofSHJG8AUucW1NJ/Calendar-Workflow?node-id=1181-13545&t=I9ovmuWndEeeXoCe-0

Update the color for dark and the color theme for the grab handle used
in the grab handle. Figma link to specific page: [Hover states of event
cards](https://www.figma.com/design/VS7WavgofSHJG8AUucW1NJ/Calendar-Workflow?node-id=2726-26013&t=I9ovmuWndEeeXoCe-0).
NOTE: For the grab handle, not all the hover states are required for the
initial version, so the colors will be added in a follow-up PRs.

## 👩‍💻 Implementation

1. Update the grab handle's (ring) dark and color theme color to
`PowerGreen`.

![image](https://github.com/user-attachments/assets/310e6275-e062-4e48-be77-f61851afb4eb)
2. Add `calendarGridBorderColor` token for all the grid border colors in
the calendar for all three themes as per the colors mentioned in the
mockup for all three themes.
3. Add `calendarGroupHeaderBackgroundColor` token for the background
color of the resource or row group header in the calendar. The dark and
color theme requires transparency so use `hexToRgbaCssColor` and set
transparency to 10%.

![image](https://github.com/user-attachments/assets/330a5eb5-1522-4a62-955c-14502a0b63e1)


## 🧪 Testing

Update the existing calendar story to add the new tokens and manually
validate the colors by inspecting.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.